pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["morcux <mircusss69@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.10.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
